+++ changelog.md
@@ -1,3 +1,7 @@
+1.2.0
+  - added method to retrieve tracker by account id and name
+  - adjusted tracking period names
+
 1.1.0
   - add more dd tracking periods